Add getItem function to Items controller
An getItem function has been added to the Items controller in order to retrieve specific items. Changes were also made to refactor the error handling in various controller functions and add a route for this new function in the Items router. A getItemBodySchema has also been added to validate the request parameters for the getItem function.