Add a seed script to populate a database with test double data
A new seed script has been added. This script populates the MongoDB database with test double data for categories, subcategories, and items. The seed script takes a JSON file of data, parses it, and feeds it into the database. Additionally, a new npm script has been added to the package.json to run the seeding operation conveniently.